refactor(modules): split things into different files
